+++ config.js
@@ -1,0 +1,7 @@
+export const factoryAddress =  "0x6337350924AdABd0345dAD410f8Eb60b2f2d9162"
+
+const factoryInitCode = "df5bb4bfdc0d140d60cf2dd372a6834035d470100d523a466e4e9067808ebb02"
+
+const router01Adress = "0x745b2d2982122741cAf1f4107edA3c15487D7B18"
+
+const routerMainAddress = "0x6A049F033a4245A30EFd30e149A302759C3fEe56"
