+++ contracts/PancakePair.sol
@@ -2,9 +2,8 @@ pragma solidity =0.5.16;
 
 import "./interfaces/IPancakePair.sol";
 import "./PancakeERC20.sol";
-// import './libraries/Math.sol';
-import "@uniswap/v2-core/contracts/libraries/Math.sol";
-import "@uniswap/v2-core/contracts/libraries/UQ112x112.sol";
+import "./libraries/Math.sol";
+import "./libraries/UQ112x112.sol";
 import "./interfaces/IERC20.sol";
 import "./interfaces/IPancakeFactory.sol";
 import "./interfaces/IPancakeCallee.sol";
