+++ contracts/interfaces/IPancakeCallee.sol
@@ -1,5 +1,10 @@
 pragma solidity >=0.5.0;
 
 interface IPancakeCallee {
-    function pancakeCall(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function pancakeCall(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
